% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_comparison.R
\name{pathway_comparision}
\alias{pathway_comparision}
\title{Projection and comparison between pathways}
\usage{
pathway_comparision(profile1,profile2,pathway_info)
}
\arguments{
\item{profile1}{A dataframe that generated by function pathway_vectorization().}

\item{profile2}{A dataframe that generated by function pathway_vectorization().}

\item{pathway_info}{A list containing pathway information to be compared, which can be generated by the merge_pathway function.}
}
\value{
Return a dataframe containing pathway ID and p-value.
}
\description{
This function makes projection and comparison between pathways. The output contains pathway ID and p-value. P-value<0.05 was considered significantly different.
}
\examples{
data(example)
significant_pathway=pathway_comparision(profile1,profile2,pathway_info)

}
